server/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
requests
openai
groq
slowapi
